make ParseRequest a top-level function
This makes it clear that parsing a request doesn't have any side effects.
It allows running a proxy that accepts CONNECT-UDP requests using
multiple different templates.
