+++ proxy_test.go
@@ -43,66 +43,6 @@ var _ http3.HTTPStreamer = &http3ResponseWriter{}
 
 func (s *http3ResponseWriter) HTTPStream() http3.Stream { return s.str }
 
-func TestUpgradeFailures(t *testing.T) {
-	s := Proxy{
-		Template: uritemplate.MustNew("https://localhost:1234/masque?h={target_host}&p={target_port}"),
-	}
-
-	t.Run("wrong request method", func(t *testing.T) {
-		req := newRequest("https://localhost:1234/masque")
-		req.Method = http.MethodHead
-		_, err := s.ParseRequest(req)
-		require.EqualError(t, err, "expected CONNECT request, got HEAD")
-		require.Equal(t, http.StatusMethodNotAllowed, err.(*RequestParseError).HTTPStatus)
-	})
-
-	t.Run("wrong protocol", func(t *testing.T) {
-		req := newRequest("https://localhost:1234/masque")
-		req.Proto = "not-connect-udp"
-		_, err := s.ParseRequest(req)
-		require.EqualError(t, err, "unexpected protocol: not-connect-udp")
-		require.Equal(t, http.StatusNotImplemented, err.(*RequestParseError).HTTPStatus)
-	})
-
-	t.Run("missing Capsule-Protocol header", func(t *testing.T) {
-		req := newRequest("https://localhost:1234/masque")
-		req.Header.Del("Capsule-Protocol")
-		_, err := s.ParseRequest(req)
-		require.EqualError(t, err, "missing Capsule-Protocol header")
-		require.Equal(t, http.StatusBadRequest, err.(*RequestParseError).HTTPStatus)
-	})
-
-	t.Run("invalid Capsule-Protocol header", func(t *testing.T) {
-		req := newRequest("https://localhost:1234/masque")
-		req.Header.Set("Capsule-Protocol", "🤡")
-		_, err := s.ParseRequest(req)
-		require.EqualError(t, err, "invalid capsule header value: [🤡]")
-		require.Equal(t, http.StatusBadRequest, err.(*RequestParseError).HTTPStatus)
-	})
-
-	t.Run("invalid Capsule-Protocol header value", func(t *testing.T) {
-		req := newRequest("https://localhost:1234/masque")
-		req.Header.Set("Capsule-Protocol", "2")
-		_, err := s.ParseRequest(req)
-		require.EqualError(t, err, "incorrect capsule header value: 2")
-		require.Equal(t, http.StatusBadRequest, err.(*RequestParseError).HTTPStatus)
-	})
-
-	t.Run("missing target host", func(t *testing.T) {
-		req := newRequest("https://localhost:1234/masque?h=&p=1234")
-		_, err := s.ParseRequest(req)
-		require.EqualError(t, err, "expected target_host and target_port")
-		require.Equal(t, http.StatusBadRequest, err.(*RequestParseError).HTTPStatus)
-	})
-
-	t.Run("invalid target port", func(t *testing.T) {
-		req := newRequest("https://localhost:1234/masque?h=localhost&p=foobar")
-		_, err := s.ParseRequest(req)
-		require.ErrorContains(t, err, "failed to decode target_port")
-		require.Equal(t, http.StatusBadRequest, err.(*RequestParseError).HTTPStatus)
-	})
-}
-
 func TestProxyCloseProxiedConn(t *testing.T) {
 	testDone := make(chan struct{})
 	defer close(testDone)
@@ -110,9 +50,7 @@ func TestProxyCloseProxiedConn(t *testing.T) {
 	remoteServerConn, err := net.ListenUDP("udp", &net.UDPAddr{IP: net.IPv4(127, 0, 0, 1), Port: 0})
 	require.NoError(t, err)
 
-	s := Proxy{
-		Template: uritemplate.MustNew("https://localhost:1234/masque?h={target_host}&p={target_port}"),
-	}
+	s := Proxy{}
 	req := newRequest(fmt.Sprintf("https://localhost:1234/masque?h=localhost&p=%d", remoteServerConn.LocalAddr().(*net.UDPAddr).Port))
 	rec := httptest.NewRecorder()
 	done := make(chan struct{})
@@ -135,7 +73,7 @@ func TestProxyCloseProxiedConn(t *testing.T) {
 		<-closeStream
 		return 0, io.EOF
 	})
-	r, err := s.ParseRequest(req)
+	r, err := ParseRequest(req, uritemplate.MustNew("https://localhost:1234/masque?h={target_host}&p={target_port}"))
 	require.NoError(t, err)
 	go s.Proxy(&http3ResponseWriter{ResponseWriter: rec, str: str}, r)
 	require.Equal(t, http.StatusOK, rec.Code)
@@ -159,11 +97,9 @@ func TestProxyCloseProxiedConn(t *testing.T) {
 }
 
 func TestProxyDialFailure(t *testing.T) {
-	s := Proxy{
-		Template: uritemplate.MustNew("https://localhost:1234/masque?h={target_host}&p={target_port}"),
-	}
+	s := Proxy{}
 	r := newRequest("https://localhost:1234/masque?h=localhost&p=70000") // invalid port number
-	req, err := s.ParseRequest(r)
+	req, err := ParseRequest(r, uritemplate.MustNew("https://localhost:1234/masque?h={target_host}&p={target_port}"))
 	require.NoError(t, err)
 	rec := httptest.NewRecorder()
